refactor: split service internals behind facades
